DashboardController is splitted #41
