+++ core/src/main/java/com/hlag/oversigt/web/resources/EventSourceStatusResource.java
@@ -20,7 +20,7 @@ import org.slf4j.LoggerFactory;
 
 import com.google.inject.Inject;
 import com.google.inject.Singleton;
-import com.hlag.oversigt.controller.DashboardController;
+import com.hlag.oversigt.controller.EventSourceInstanceController;
 import com.hlag.oversigt.core.eventsource.EventSourceStatisticsManager;
 import com.hlag.oversigt.core.eventsource.EventSourceStatisticsManager.EventSourceStatistics;
 import com.hlag.oversigt.core.eventsource.EventSourceStatisticsManager.RunStatistic;
@@ -46,7 +46,7 @@ public class EventSourceStatusResource {
 	private static final Logger LOGGER = LoggerFactory.getLogger(EventSourceStatusResource.class);
 
 	@Inject
-	private DashboardController controller;
+	private EventSourceInstanceController controller;
 
 	@Inject
 	private EventSourceStatisticsManager statisticsManager;
@@ -114,7 +114,7 @@ public class EventSourceStatusResource {
 	}
 
 	public static class EventSourceInstanceState {
-		public static EventSourceInstanceState fromInstance(final DashboardController controller,
+		public static EventSourceInstanceState fromInstance(final EventSourceInstanceController controller,
 				final EventSourceStatisticsManager statisticsManager,
 				final EventSourceInstance instance) {
 			final EventSourceStatistics stats = statisticsManager.getEventSourceStatistics(instance.getId());
